llo/common: extract proto conversions
